diagnose/from django.py
from django.db import models

class Todo(models.Model):
    title = models.CharField(max_length=200)
    description = models.TextField()

    def __str__(self):
        return self.title

#############################################################################

python manage.py migrate

#############################################################################

from django.contrib.auth.models import Group, User
from rest_framework import serializers

class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ['url', 'username', 'email', 'groups']

class GroupSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Group
        fields = ['url', 'name']

#############################################################################

from django.contrib.auth.models import User
from rest_framework import permissions, viewsets
from tutorial.quickstart.serializers import GroupSerializer, UserSerializer

class UserViewSet(viewsets.ModelViewSet):
    queryset = User.objects.all().order_by('-date_joined')
    serializer_class = UserSerializer

###############################################################################
Translate this Python code into Java
Import operating system
Import CV2
Import numpy as np
Import Tensorflow as tf

disease = {0:"normal", 1:"COVID-19", 2:"viral pneumonia", 3:"lung opacification", 4:"bacterial pneumonia"}

#Current_file_path
cur = os.path.abspath(__file__).replace('\\', '/')

model_path = cur[:-11] + '/all_diseases_2.h5'
Model = tf.keras.models.load_model(model_path)


Definitely expect (img):
     images = []
     If img is none:
         Returns "nil from predicate"
     image=image/255.0.0
     img = cv2.resize(img, (100, 100))
     Append images (img)
     img = np.asarray(images)
     y_pred = model.predict(img)
     Return disease[np.argmax(y_pred[0])]

from django.urls import path, include
from rest_framework import routers
from tutorial.quickstart.views import UserViewSet

router = routers.DefaultRouter()
router.register(r'users', UserViewSet)

urlpatterns = [
    path('', include(router.urls)),
]


